requirements.txt
yt-dlp>=2023.11.16
pyyaml>=6.0.1
pydantic>=2.5.2
loguru>=0.7.2
faster-whisper>=1.0.0
torch>=2.0.0
openai>=1.0.0
anthropic>=0.18.0
instructor>=1.0.0
tiktoken>=0.6.0
mediapipe>=0.10.0
opencv-python>=4.9.0
numpy>=1.26.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pillow>=10.0.0
moviepy>=1.0.3
scikit-image>=0.22.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
playwright>=1.40.0
